Rename software to Ard
Changed from windArd to Ard due to legal and copyright requirements
