+++ .gitignore
@@ -1,5 +1,5 @@
 
-### WINDARD DEVELOPMENT IGNORES 
+### ARD DEVELOPMENT IGNORES 
 
 case_files/
