+++ examples/FLORIS_power_comp/power_comp.py
@@ -8,9 +8,9 @@ import openmdao.api as om
 
 from wisdem.inputs.validation import load_yaml
 
-import windard.utils
-import windard.wind_query as wq
-import windard.farm_aero.floris as farmaero_floris
+import ard.utils
+import ard.wind_query as wq
+import ard.farm_aero.floris as farmaero_floris
 
 # create the wind query
 directions = np.linspace(0.0, 360.0, 101)
@@ -48,7 +48,7 @@ filename_floris_config = os.path.abspath(
     )
 )  # default FLORIS config for the project
 # create a FLORIS yaml to conform to the config/spec files above
-windard.utils.create_FLORIS_yamlfile(filename_turbine_spec, filename_turbine_FLORIS)
+ard.utils.create_FLORIS_yamlfile(filename_turbine_spec, filename_turbine_FLORIS)
 # load the turbine specification
 data_turbine = load_yaml(filename_turbine_spec)
