+++ pyproject.toml
@@ -1,5 +1,5 @@
 
-# Project configuration for windArd
+# Project configuration for Ard
 # TOML spec: https://toml.io/en/v1.0.0
 
 [build-system]
@@ -10,10 +10,10 @@ build-backend = "setuptools.build_meta"
 
 # https://setuptools.pypa.io/en/latest/userguide/package_discovery.html
 [tool.setuptools]
-packages = ["windard"]
+packages = ["ard"]
 
 [project]
-name = "windArd"
+name = "Ard"
 version = "0.0.0a1"
 authors = [
   {name = "Cory Frontin", email = "cory.frontin@nrel.gov"},
@@ -51,7 +51,7 @@ docs = [
 ]
 [project.urls]
 # Homepage = "https://example.com"
-# Documentation = "https://wisdem.github.io/windArd"
-Repository = "https://github.com/wisdem/windArd.git"
-Issues = "https://github.com/wisdem/windArd/issues"
-# Changelog = "https://github.com/wisdem/windArd/blob/main/CHANGELOG.md"
+# Documentation = "https://wisdem.github.io/Ard"
+Repository = "https://github.com/wisdem/Ard.git"
+Issues = "https://github.com/wisdem/Ard/issues"
+# Changelog = "https://github.com/wisdem/Ard/blob/main/CHANGELOG.md"
