+++ test/run_local_test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-# pytest --cov=windard test/unit/cost
-pytest --cov=windard test
+# pytest --cov=ard test/unit/cost
+pytest --cov=ard test
 
 rm -rf test/unit/layout/problem*_out
